+++ packages/document/template/feedback/popconfirm-disabled.vue
@@ -1,7 +1,5 @@
 <template>
-	<px-space>
-		<px-popconfirm content="This popconfirm is disabled" disabled>
-			<px-button>Disabled</px-button>
-		</px-popconfirm>
-	</px-space>
+	<px-popconfirm content="This popconfirm is disabled" disabled>
+		<px-button>Disabled</px-button>
+	</px-popconfirm>
 </template>