+++ packages/document/template/navigation/menu-options.vue
@@ -1,0 +1,59 @@
+<template>
+	<px-menu :options="options"> </px-menu>
+</template>
+
+<script setup lang="ts">
+import {
+	IconHome,
+	IconArrowAltCircleUp,
+	IconPencil,
+	IconGithub,
+	IconLink,
+	IconBars
+} from '@pixelium/web-vue/icon-hn/es'
+import { h, ref } from 'vue'
+
+const options = ref([
+	{
+		label: 'Home',
+		index: 'Home',
+		href: '/pixelium-design',
+		icon: h(IconHome)
+	},
+	{
+		label: 'Components',
+		index: 'Components',
+		icon: h(IconBars),
+		type: 'submenu',
+		children: [
+			{
+				label: 'Button',
+				index: 'Button',
+				icon: h(IconArrowAltCircleUp)
+			},
+			{
+				label: 'Input',
+				index: 'Input',
+				icon: h(IconPencil)
+			}
+		]
+	},
+	{
+		label: 'More',
+		index: 'More',
+		type: 'group',
+		children: [
+			{
+				label: 'About',
+				index: 'About',
+				icon: h(IconLink)
+			},
+			{
+				label: 'Github',
+				index: 'Github',
+				icon: h(IconGithub)
+			}
+		]
+	}
+])
+</script>