+++ packages/web-vue/lib/button-group/index.vue
@@ -22,7 +22,6 @@ defineOptions({
 })
 
 const props = withDefaults(defineProps<ButtonGroupProps>(), {
-	shape: 'rect',
 	disabled: false
 })