+++ packages/web-vue/lib/col/index.vue
@@ -19,7 +19,7 @@
 import { isNumber } from 'parsnip-kit'
 import { useScreenWidth } from '../share/hook/use-screen-width'
 import type { ColProps } from './type'
-import { computed, getCurrentInstance, inject, ref, type Ref } from 'vue'
+import { computed, inject, type Ref } from 'vue'
 
 defineOptions({
 	name: 'Col'
@@ -44,13 +44,7 @@ const spanComputed = computed(() => {
 	return props.span[widthType.value] || 24
 })
 
-const instance = getCurrentInstance()
-const inner = ref(instance?.parent?.type.name === 'Row')
-
-const gutter =
-	(inner.value &&
-		inject<Ref<{ x: number; y: number }> | undefined>('px-row-provide', undefined)) ||
-	ref({ x: 0, y: 0 })
+const gutter = inject<Ref<{ x: number; y: number }> | undefined>('px-row-provide', undefined)
 </script>
 <style lang="less" src="./index.less"></style>
 <style lang="less" src="../share/style/index.css" />