+++ packages/web-vue/lib/popover/index.vue
@@ -1,5 +1,5 @@
 <script setup lang="tsx">
-import { useSlots, shallowRef, useAttrs } from 'vue'
+import { useSlots, shallowRef, mergeProps } from 'vue'
 import Popup from '../popup/index.vue'
 import type { PopoverProps, PopoverEvents } from './type'
 import { forwardEmits } from '../share/util/reactivity'
@@ -21,7 +21,6 @@ const props = withDefaults(defineProps<PopoverProps>(), {
 	destroyOnHide: false
 })
 
-const attrs = useAttrs()
 const emits = defineEmits<PopoverEvents>()
 
 const popupRef = shallowRef<InstanceType<typeof Popup> | null>(null)
@@ -41,7 +40,7 @@ const forward = forwardEmits(emits, ['open', 'close', 'update:visible'])
 
 defineRender(() => {
 	return (
-		<Popup ref={popupRef} {...{ ...props, ...forward, ...attrs }}>
+		<Popup ref={popupRef} {...mergeProps(props, forward)}>
 			{{
 				default: slots.default,
 				content: slots.content