+++ packages/web-vue/lib/share/const/index.ts
@@ -39,3 +39,5 @@ export const SQRT3 = Math.sqrt(3)
 export const INV_SQRT3 = 1 / SQRT3
 
 export const GET_ELEMENT_RENDERED = 'getElementRendered'
+
+export const SUBMENU_OPTION_TYPE = 'submenu'